requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
braceexpand==0.1.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
isort==5.12.0
jishaku==2.5.1
line-profiler==4.0.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
python-dotenv==1.0.0
six==1.16.0
yarl==1.9.1
